requirements.txt
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.7.14
click==8.2.1
clip @ git+https://github.com/openai/CLIP.git@dcba3cb2e2827b402d2701e7e1c7d9fed8a20ef1
databases==0.9.0
distro==1.9.0
fastapi==0.116.1
filelock==3.13.1
fsspec==2024.6.1
ftfy==6.3.1
greenlet==3.2.3
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
Jinja2==3.1.4
jiter==0.10.0
loguru==0.7.3
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.3
numpy==2.2.6
openai==1.98.0
opencv-python==4.12.0.88
orm==0.3.1
packaging==25.0
pillow==11.3.0
pydantic==2.11.7
pydantic_core==2.33.2
python-multipart==0.0.20
regex==2025.7.34
setuptools==70.2.0
sniffio==1.3.1
SQLAlchemy==2.0.42
starlette==0.47.2
sympy==1.13.3
tqdm==4.67.1
typesystem==0.3.1
typing-inspection==0.4.1
typing_extensions==4.14.1
uvicorn==0.35.0
wcwidth==0.2.13
